style(homepage): 어바웃어스 섹션별 스크롤 스냅 적용 및 진입 애니메이션 최적화 (#280)
